backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
pydantic-settings
python-dotenv
protobuf
grpcio
google-api-python-client

# Database Drivers
neo4j
chromadb-client

# Optional: If doing embedding generation within the API (not recommended for heavy load)
# sentence-transformers

# Optional: For actual synthesis generation
openai
langchain # Or other relevant LLM libraries 

# Added from the code block
langchain-community
langchain-openai
click 